One conflict, in `vector_score_source::build_result`: this branch added the
conservative `result.has_field_expiration = true` where master's MOD-14210
(#10221) replaced the same TODO with a real metrics push and added a new
`attach_score_metric` trait method.

Kept both. The forced bit stays in `build_result` only, which is the unfiltered
path where the source fabricates a result with no inverted-index block behind
it. `attach_score_metric` is left as master wrote it: there the yielded record
is the child's, which already carries the accurate per-entry bit from the
block bitset, so forcing it would only discard the optimization.

+++ src/redisearch_rs/c_entrypoint/inverted_index_ffi/src/fork_gc.rs
@@ -89,7 +89,7 @@ mod tests {
     extern "C" fn vec_writer(ctx: *mut c_void, buf: *const c_void, len: size_t) {
         unsafe {
             let v = &mut *(ctx as *mut Vec<u8>);
-            let src = core::slice::from_raw_parts(buf as *const u8, len as usize);
+            let src = core::slice::from_raw_parts(buf as *const u8, len);
             v.extend_from_slice(src);
         }
     }
@@ -97,7 +97,7 @@ mod tests {
     extern "C" fn vec_reader(ctx: *mut c_void, buf: *mut c_void, len: size_t) -> c_int {
         unsafe {
             let v = &mut *(ctx as *mut Vec<u8>);
-            let want = len as usize;
+            let want = len;
             if v.len() < want {
                 return 1;
             }
